🧪 test(postgres): Add comprehensive integration tests for repositories
This commit introduces a suite of new integration tests for various
Postgres repositories within the `core` module. These tests are
essential for ensuring the correct functionality, data integrity, and
persistence behavior of the Rust repository implementations. By
mirroring existing tests from the `legacy/storage/gorm` package, we
ensure consistency and validate the Rust versions against established
expectations.

The addition of these tests significantly boosts the confidence in the
Postgres repository layer, verifying operations such as saving,
retrieving, updating, and querying records for different domain
entities.

**Changes made:**
- **Added** `AuditLogRepository` integration tests in
  `core/src/repository/postgres/audit_log_test.rs`. These tests cover
  saving and retrieving audit log entries.
- **Implemented** `FeedbackRecordRepository` tests in
  `core/src/repository/postgres/feedback_test.rs`, validating feedback
  submission and retrieval.
- **Created** `GroupRepository` tests in
  `core/src/repository/postgres/group_test.rs` to ensure group creation,
  retrieval, and updates function as expected.
- **Introduced** `MatchRepository` tests in
  `core/src/repository/postgres/match_repo_test.rs`, covering complex
  match-related operations like creation, status updates, and retrieval
  by various criteria.
- **Extended** test coverage to `OfferRepository`, `RawMessageRepository`,
  `RequestRepository`, `ReviewQueueRepository`, `StatsRepository`, and
  `WeightHistoryRepository` with dedicated `_test.rs` files. Each file
  provides focused integration tests for its respective repository.
- **Integrated** `TestDb` and helper functions such as
  `new_test_audit_log`, `new_test_feedback_record`, etc., within the
  `core/src/repository/postgres/testing.rs` module (though not fully
  shown in diff, its usage is pervasive), ensuring consistent and
  reproducible test setups.
- **Mirrored** the logic and test scenarios found in the
  `legacy/storage/gorm` repository tests to provide a direct comparison
  and validation against the previous Go implementation.

> **Note:** The new test files leverage `#[tokio::test]` for asynchronous
> execution, making them suitable for testing database interactions.

**Technical details:**
-   **Files added:**
    -   `core/src/repository/postgres/audit_log_test.rs`
    -   `core/src/repository/postgres/feedback_test.rs`
    -   `core/src/repository/postgres/group_test.rs`
    -   `core/src/repository/postgres/match_repo_test.rs`
    -   `core/src/repository/postgres/offer_test.rs`
    -   `core/src/repository/postgres/raw_message_test.rs`
    -   `core/src/repository/postgres/request_test.rs`
    -   `core/src/repository/postgres/review_queue_test.rs`
    -   `core/src/repository/postgres/stats_test.rs`
    -   `core/src/repository/postgres/weight_history_test.rs`
    -   `core/src/repository/postgres/mod.rs` (likely updated to include new modules)
    -   `core/src/repository/postgres/stats.rs` (likely updated/extended)
    -   `core/src/repository/postgres/testing.rs` (likely extended with new test helpers)
-   **Functions introduced (within tests):** `test_save`, `test_get_all`,
    `test_get_by_id`, `test_get_page`, `test_update`, `test_delete`,
    `test_get_match_between`, `test_get_by_request_id`, etc., tailored
    to each repository's functionality.
-   **Dependencies:** `chrono` for time-based assertions, `uuid` for
    ID generation, `tokio` for async test runtime.

+++ core/src/repository/postgres/stats.rs
@@ -49,7 +49,7 @@ impl StatsRepository for PostgresStatsRepo {
         let monitored_groups = sqlx::query("SELECT COUNT(*) FROM groups WHERE monitored = true")
             .fetch_one(&self.pool)
             .await?
-            .get::<i32, _>(0);
+            .get::<i64, _>(0) as i32;
 
         Ok(Stats {
             active_offers,
